requirements.txt
# requirements.txt
slack-bolt==1.18.0
slack-sdk==3.23.0
python-dotenv==1.0.0
flask==3.0.0

# Database & Cache
supabase==2.0.3
redis==5.0.1

# ML & Data Processing
sentence-transformers==3.0.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.0

# Web Scraping & Search
requests==2.31.0
beautifulsoup4==4.12.2

# LLM
groq==0.4.1

# PDF Generation
reportlab==4.0.7

# Email
sendgrid==6.11.0

# Utilities
python-slugify==8.0.1